+++ Sources/TuistCoreTesting/Utils/MockXcodeProjectBuildDirectoryLocator.swift
@@ -1,0 +1,19 @@
+import TSCBasic
+import TuistCore
+import TuistGraph
+
+public final class MockXcodeProjectBuildDirectoryLocator: XcodeProjectBuildDirectoryLocating {
+    public init() {}
+
+    enum MockXcodeProjectBuildDirectoryLocatorError: Error {
+        case noStub
+    }
+
+    public var locateStub: ((Platform, AbsolutePath, String) throws -> AbsolutePath)?
+    public func locate(platform: Platform, projectPath: AbsolutePath, configuration: String) throws -> AbsolutePath {
+        guard let stub = locateStub else {
+            throw MockXcodeProjectBuildDirectoryLocatorError.noStub
+        }
+        return try stub(platform, projectPath, configuration)
+    }
+}