+++ Sources/TuistKit/Commands/BuildCommand.swift
@@ -40,6 +40,12 @@ struct BuildCommand: ParsableCommand {
     )
     var configuration: String?
 
+    @Option(
+        help: "The directory where build products will be copied to when the project is built.",
+        completion: .directory
+    )
+    var buildOutputPath: String?
+
     func run() throws {
         let absolutePath: AbsolutePath
         if let path = path {
@@ -47,11 +53,13 @@ struct BuildCommand: ParsableCommand {
         } else {
             absolutePath = FileHandler.shared.currentPath
         }
+
         try BuildService().run(
             schemeName: scheme,
             generate: generate,
             clean: clean,
             configuration: configuration,
+            buildOutputPath: buildOutputPath.map { AbsolutePath($0, relativeTo: FileHandler.shared.currentPath) },
             path: absolutePath
         )
     }